added proifile controller and profile routes besides the profile component
